clean up status table before proceeding with migration
